# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UniCebim API
  version: 0.1.0
  description: |
    UniCebim (Student Budget Tracker) MVP API specification.

    Notes:
    - Auth is handled via Supabase session cookies in the web app.
    - Endpoints below assume an authenticated session for protected routes.
servers:
  - url: http://localhost:3000

tags:
  - name: Health
  - name: Profile
  - name: Transactions

paths:
  /api/health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok, service]
                properties:
                  ok:
                    type: boolean
                  service:
                    type: string

  /api/profile:
    get:
      tags: [Profile]
      summary: Get current user's profile
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProfileResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [Profile]
      summary: Update current user's monthly budget goal
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMonthlyBudgetGoalRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProfileResponse"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/transactions:
    get:
      tags: [Transactions]
      summary: List transactions for the current user (optional month filter)
      security:
        - cookieAuth: []
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}$"
          description: Month filter in YYYY-MM format (UTC month boundaries).
      responses:
        "200":
          description: Transactions + monthly summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTransactionsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Transactions]
      summary: Create a transaction for the current user
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [transaction]
                properties:
                  transaction:
                    $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token

  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    TransactionType:
      type: string
      enum: [income, expense]

    Transaction:
      type: object
      required: [id, amount, type, category, date]
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
        type:
          $ref: "#/components/schemas/TransactionType"
        category:
          type: string
        date:
          type: string
          format: date-time

    MonthlySummary:
      type: object
      required: [incomeTotal, expenseTotal, netTotal]
      properties:
        incomeTotal:
          type: number
        expenseTotal:
          type: number
        netTotal:
          type: number

    ListTransactionsResponse:
      type: object
      required: [month, summary, transactions]
      properties:
        month:
          type: string
          pattern: "^\\d{4}-\\d{2}$"
        summary:
          $ref: "#/components/schemas/MonthlySummary"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"

    CreateTransactionRequest:
      type: object
      required: [amount, type, category, date]
      properties:
        amount:
          type: number
        type:
          $ref: "#/components/schemas/TransactionType"
        category:
          type: string
        date:
          type: string
          format: date
          description: Date-only string (YYYY-MM-DD).

    Profile:
      type: object
      required: [id, full_name, monthly_budget_goal]
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
          nullable: true
        monthly_budget_goal:
          type: number
          nullable: true

    GetProfileResponse:
      type: object
      required: [profile]
      properties:
        profile:
          $ref: "#/components/schemas/Profile"
          nullable: true

    UpdateMonthlyBudgetGoalRequest:
      type: object
      properties:
        monthlyBudgetGoal:
          type: number
          nullable: true